Use updated `Game` constructor path searching
Rather than specify a relative path from the `demoGraphics` project folder, use the path mount searching capability of the `Game` object to find the appropriate `data/` folder.

There is no `data/` folder in the NAS2D repository, though if it is a submodule of a larger repository, such as OPHD, then it may find a `data/` folder, which would have the desired `Font` files.