Return newly constructed value in creation block
If we update how objects are created to allow for factory functions, then we should perhaps not assume an equivalent iterator will be available, or is the best way to get a reference to the new object. Maybe better to assume we can directly return a reference to the newly created object.