Use `elapsedTicks` instead of `ticks` in DemoGraphics project
The `ticks` method is a global/static function, which returns a value unrelated to the specific `Timer` object. If a delay is added to startup, such as by `sleep(2)`, then using `ticks` will cause the gears to be already significantly rotated when they come into view. Whereas using `elapsedTicks` will give a value relative to when the `Timer` was created, which means the gears will be in their initial start position.