Add `AnimationSequence` destructor
This is needed to force instantiation of the `std::vector` destructor in the implementation file, rather than in the translation unit that includes the `AnimationSequence` header. This is important since the element type `AnimationFrame` is forward declared in the header, and a full definition is needed to instantiate the `std::vector` destructor.