Initialize `Sprite` fields in constructor init list
Now that `Sprite` has only the single constructor overload, it makes sense to initialize the fields there.

Running `cppclean` was crashing on the default value for the `bool` field using the `{false}` initializer syntax. Switching to the `=` syntax worked fine for `cppclean`. Both are perfectly valid syntax, and there are other fields that were using the initializer syntax. Of course now that there is only one constructor overload, using the constructor init list is preferred to initializing fields in the header file.

Keep the empty initializer for `Timer` field since otherwise we get a warning from `-Weffc++` with GCC.