Fix error reporting of bad `version` in sprite file loading
This bug has some interesting history. Back in commit 6738da544acf78c22ba84b549bcf0982a15aa3a3 a local variable named `versionString` was removed, which was shadowing the name of a global function, while the error reporting line were the variable was used was not updated. This mean the error reporting line was now printing the address of the global function rather than the version string read from the file.

Later on, in commit 585b734ae4dd42a072f7557a4a1ca1346e65433c it was noticed from a warning message that a function address was being printed. This was "corrected" by calling the function and passing the result as part of the error reporting message, rather than printing the address of the global function. It wasn't noticed the function didn't have anything to do with the value that should have been printed. And why would it be noticed, since this was now 4 months after the removal of the shadowing local variable by a different committer. There was no obvious context, and it was assumed the error was simply a lack of parenthesis to actually call the function, rather than the function not being the original intended target for printing.