Use `struct` instead of `using` alias for path types
This prevents swapping path types in API calls. With a `using` alias there is no new type created, so anything aliased to the same type will be interchangeable with each other. By using `struct` we create new types, and so are forced to use the exact type, rather than an equivalent name. As such we now get warnings if a `RealPath` is passed where a `VirtualPath` is expected, or a `VirtualPath` is passed where a `RealPath` is expected.