Decouple `AnimationSequence` loading from `ResourceCache`
Use a `Delegate` to wrap the image load method, allowing other sources to be used, and to isolate `AnimationFile` from the details of `ResourceCache`. In particular, `AnimationFile` no longer depends on a specific template instantiation of `ResourceCache`, which may matter if factory functions get included in the `ResourceCache` type information.