Use `std::forward` in `Signal::operator()`
This reduces copying of rvalue parameters when passing them from `operator()` to `emit`. Note that `emit` will still need to copy parameters, since it must pass them on to potentially multiple listeners. Moving the values would likely invalidate the source copy, so could only be used if there was a single listener.